fix:sonar issues
